+++ Editor/ARCameraFeatureEditor.cs
@@ -1,0 +1,11 @@
+using UnityEngine;
+using UnityEngine.XR.OpenXR.Features.Meta;
+
+namespace UnityEditor.XR.OpenXR.Features.Meta
+{
+    [CustomEditor(typeof(ARCameraFeature))]
+    class ARCameraFeatureEditor : OpenXRFeatureEditorBase
+    {
+
+    }
+}