+++ Tests/Runtime/NativeTypes/Core/XrGetSystemProperties_Delegate.cs
@@ -1,5 +1,5 @@
 namespace UnityEngine.XR.OpenXR.Features.Meta.Tests
 {
-    unsafe delegate void GetSystemProperties_Delegate(
+    unsafe delegate void GetSystemProperties_delegate(
         XrSystemPropertiesBaseHeader* systemPropertiesStruct);
 }